Changes requested by @femtobit and @gcarleo during review
