+++ Test/Machine/rbm.py
@@ -1,0 +1,1 @@
+../../Examples/CustomMachine/rbm.py
